requirements.txt
Django==5.1.6
psycopg2-binary==2.9.10
django-environ==0.12.0
djangorestframework == 3.15.2
django-bootstrap-v5 == 1.0.11